requirements.txt
asgiref==3.2.10
coverage==5.3
Django==3.1.2
djangorestframework==3.12.1
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.4.1
